common-videooutput: store the extended stream buffer
When a second remote video stream (extended stream) arrives, it is
stored in a new byte array called eframeStore. The dimensions of
the frame are also store, along with a new boolean property that
confirms the existance of an extended stream in the connection.

This patch doesn't modifies the current user experience, but it is
a requirement for the extended video display.