x-videooutput: resources for the extended stream
This patch adds the X resources required to display the extended stream in
case the user selects that view.

The resources are a XWindow and a Display instances.

The patch contains the logic to handle these two new instances: construction,
destruction, and also receives the frame from the byte array and the display
synchronization.

Since there is no a way in the user interface for enabling the extended stream
view, this patch doesn't modify the user experience.

TODO: add this bits for the dx-videooutput. Since I don't have any experience,
nor access to DirectX, I prefer avoid its modification.