common-videooutput: extended stream in frame info
This patch adds a boolean variable (ext_stream_active). It tells if there is
an extended remote video stream in the connection. It is stored in the frame
info (FrameInfo) structure.

FrameInfo is the structure that is going to be used for each frame that
arrives. So if will tell to the user interface if connection contains an
extended video.

The patch specifies the value of this new attribute for current_frame and
last_frame instances of FrameInfo.