common-videooutput: display extended stream
This patch contains the logic when the extended remote video stream changes
its size, hence the UI window should be adjusted.

Also contains the case that if the stream adds or removes the extended stream,
the UI setup should be readjusted too.

Finally this patch handles the frame display of the extended remote stream if
the UI requests it.

No UI changes yet though.